add compare apps command
