+++ skills/cli/e2e/compare_sql.json
